requirements.txt
flask
python-dotenv
requests
supabase
pinecone-client
openai
boto3 